Extract extra code from SpotlightMap
