+++ components/SpotlightMap/index.ts
@@ -1,1 +1,2 @@
 export * from './SpotlightMap';
+export * from './utils/types';
